move test back in volunteer.test.ts
